Telegram branch

+++ src/main/java/ru/urgu/vkDialogueBot/Controller/SimpleUserToken.java
@@ -6,13 +6,13 @@ import lombok.Setter;
 public class SimpleUserToken implements IUserToken
 {
     @Getter
-    private final int _hash;
+    private final Long _hash;
 
     @Getter
     @Setter
     private int _currentResponderId = -1;
 
-    public SimpleUserToken(int hash)
+    public SimpleUserToken(Long hash)
     {
         _hash = hash;
     }
